More elf support
