+++ cli/cli.ts
@@ -25,7 +25,6 @@ import * as serial from './serial';
 import * as gdb from './gdb';
 import * as clidbg from './clidbg';
 import * as pyconv from './pyconv';
-import * as elf from './elf';
 
 const rimraf: (f: string, opts: any, cb: () => void) => void = require('rimraf');
 
@@ -3248,14 +3247,6 @@ function dbgTestAsync() {
         .then(clidbg.startAsync)
 }
 
-function elfAsync(parsed: commandParser.ParsedCommand) {
-    let fn = parsed.arguments[0]
-    let buf = fs.readFileSync(fn)
-    let buf2 = elf.patchElf(fn, buf as any)
-    fs.writeFileSync("a.out", buf2)
-    return Promise.resolve()
-}
-
 interface BuildCoreOptions {
     mode: BuildOption;
 
@@ -4300,7 +4291,6 @@ function initCommands() {
     advancedCommand("testconv", "test TD->TS converter", testConverterAsync, "<jsonurl>");
     advancedCommand("testpkgconflicts", "tests package conflict detection logic", testPkgConflictsAsync);
     advancedCommand("testdbg", "tests hardware debugger", dbgTestAsync);
-    advancedCommand("elf", "ELF processing", elfAsync, "<ELF-file>");
 
     advancedCommand("buildtarget", "build pxtarget.json", buildTargetAsync);
     advancedCommand("uploadtrg", "upload target release", pc => uploadTargetAsync(pc.arguments[0]), "<label>");
